+++ python/lsst/ts/criopy/gui/actuatorsdisplay/force_actuator_item.py
@@ -20,13 +20,13 @@
 
 import enum
 
-from lsst.ts.salobj import BaseMsgType
 from lsst.ts.xml.tables.m1m3 import ForceActuatorData
 from PySide6.QtCore import QPointF, QRect, Qt
-from PySide6.QtGui import QBrush, QGuiApplication, QPainter, QPalette, QPen, QTransform
-from PySide6.QtWidgets import QGraphicsItem, QStyleOptionGraphicsItem, QWidget
+from PySide6.QtGui import QGuiApplication, QPainter, QPalette, QPen
+from PySide6.QtWidgets import QStyleOptionGraphicsItem, QWidget
 
 from ...gui import Colors
+from .data_item import DataItem, DataItemState
 from .gauge_scale import GaugeScale
 
 
@@ -37,7 +37,7 @@ class FASelection(enum.IntEnum):
     FAR_NEIGHBOR = 4
 
 
-class ForceActuatorItem(QGraphicsItem):
+class ForceActuatorItem(DataItem):
     """Combines graphical display of an actuator with its data. Record if an
     actuator is selected by a mouse click.
 
@@ -51,77 +51,32 @@ class ForceActuatorItem(QGraphicsItem):
     ----------
     actuator : `ForceActuatorData`
         Row from force actuator table.
-    data : `float`
-        Data associated with the actuator (actual force, calculated force, ..).
-    data_index : `int`
-        Index in value arrays. Points to selected actuator value.
-    scale: `object`
-        Object providing getColor(value) method.
-    state : `int`
-        Force Actuator state. 0 for inactive/unused, 1 for active OK, 2 for
-        active warning.
+    state : `DataItemState`
+        Force Actuator state.
     kind : `FASelection`
         FA kind - normal, selected or neighbour of selected.
-    """
-
-    STATE_INACTIVE = 0
-    """Force Actuator value is not relevant for the current display (`int`).
-    """
-
-    STATE_ACTIVE = 1
-    """Force Actuator is active and healthy (`int`).
-    """
-
-    STATE_WARNING = 2
-    """Force Actuator is active, but the value / actuator has some warning
-    attached (`int`).
     """
 
     def __init__(
         self,
         actuator: ForceActuatorData,
-        data: BaseMsgType,
-        data_index: int | None,
-        state: int,
-        kind: FASelection,
+        state: DataItemState = DataItemState.INACTIVE,
+        kind: FASelection = FASelection.NORMAL,
     ):
-        super().__init__()
+        super().__init__(state)
         self.actuator = actuator
         # actuator position
         self._center = QPointF(
             actuator.x_position * 1000.0, -actuator.y_position * 1000.0
         )
-        # actuator data
-        self._data = data
-        self.data_index = data_index
-        self._state = state
         self._kind = kind
-        # scale. Provides getColor(data) object, returning brush to fill data
+        # scale. Provides get_brush(data) object, returning brush to fill data
         self._color_scale: None | GaugeScale = None
         # scalign factor. The actuator default size is 20x20 units. As
         # actuators are placed on mirror, the size needs to be adjusted to show
         # properly actuator on display in e.g. mm (where X and Y ranges are
         # ~-4400 .. +4400).
         self._scale_factor = 25
-
-    def update_data(self, data: float, state: int) -> None:
-        """Updates actuator data.
-
-        If new data differs from the current data, calls update() to force
-        actuator redraw.
-
-        Parameters
-        ----------
-        data : `float`
-             New data associated with the actuator (actual force, calculated
-             force, ..).
-        state : `state`
-             New actuator state value.
-        """
-        if self._data != data or self._state != state:
-            self._data = data
-            self._state = state
-            self.update()
 
     def setKind(self, kind: FASelection) -> None:
         """Set actuator kind (selection status).
@@ -140,6 +95,7 @@ class ForceActuatorItem(QGraphicsItem):
     @property
     def data(self) -> float:
         """Value associated with the actuator (`float`)."""
+        assert self._data is not None
         return self._data
 
     @data.setter
@@ -150,12 +106,12 @@ class ForceActuatorItem(QGraphicsItem):
     @property
     def warning(self) -> bool:
         """If actuator is in warning state (`bool`)."""
-        return self._state == self.STATE_WARNING
+        return self._state == DataItemState.WARNING
 
     @property
     def active(self) -> bool:
         """If actuator is active (`bool`)."""
-        return not (self._state == self.STATE_INACTIVE)
+        return not (self._state == DataItemState.INACTIVE)
 
     def set_color_scale(self, scale: GaugeScale) -> None:
         """Set actuator data display scale. This is used for setting display
@@ -164,7 +120,7 @@ class ForceActuatorItem(QGraphicsItem):
         Parameters
         ----------
         scale : `object`
-            Scaling object. Should provide formatValue() and getColor()
+            Scaling object. Should provide format_value() and get_brush()
             methods."""
         self._color_scale = scale
         self.update()
@@ -178,16 +134,16 @@ class ForceActuatorItem(QGraphicsItem):
             20 * self._scale_factor,
         )
 
-    def getValue(self) -> str:
+    def get_value(self) -> str:
         """Returns current value, string formated to scale.
 
         Returns
         -------
-        formatValue : `str`
+        format_value : `str`
            Current value formatted by the currently used color scale."""
-        return self.formatValue(self.data)
+        return self.format_value(self.data)
 
-    def formatValue(self, v: float) -> str:
+    def format_value(self, v: float) -> str:
         """Returns
 
         Parameters
@@ -203,7 +159,7 @@ class ForceActuatorItem(QGraphicsItem):
         """
         if self._color_scale is None:
             return str(v)
-        return self._color_scale.formatValue(v)
+        return self._color_scale.format_value(v)
 
     def paint(
         self, painter: QPainter, option: QStyleOptionGraphicsItem, widget: QWidget
@@ -236,7 +192,7 @@ class ForceActuatorItem(QGraphicsItem):
         painter.setRenderHint(QPainter.Antialiasing)
         painter.setRenderHint(QPainter.SmoothPixmapTransform)
         # paint grayed circle for actuators not providing the selected value
-        if self._state == self.STATE_INACTIVE:
+        if self._state == DataItemState.INACTIVE:
             painter.setPen(QPen(Qt.gray, self._scale_factor, Qt.DotLine))
             painter.drawEllipse(
                 self._center, 10 * self._scale_factor, 10 * self._scale_factor
@@ -256,23 +212,12 @@ class ForceActuatorItem(QGraphicsItem):
             painter.setPen(QPen(Qt.red, self._scale_factor, lineStyle))
 
         # draw actuator with warning in red color
-        if self._state == self.STATE_WARNING:
+        if self._state == DataItemState.WARNING:
             painter.setBrush(Colors.ERROR)
         else:
-            color = self._color_scale.getColor(self._data)
-            if color is None:
-                brush = QBrush(Qt.red, Qt.DiagCrossPattern)
-                brush.setTransform(
-                    QTransform().scale(self._scale_factor / 3, self._scale_factor / 3)
-                )
-                painter.setBrush(brush)
-            else:
-                painter.setBrush(
-                    QBrush(
-                        color,
-                        Qt.SolidPattern if self.isEnabled() else Qt.Dense4Pattern,
-                    )
-                )
+            assert self._data is not None
+            brush = self._color_scale.get_brush(self._data)
+            painter.setBrush(brush)
         # draw actuator, write value
         painter.drawEllipse(
             self._center, 10 * self._scale_factor, 10 * self._scale_factor
@@ -282,7 +227,7 @@ class ForceActuatorItem(QGraphicsItem):
 
         draw_actuator_id()
 
-        vstr = self.getValue()
+        vstr = self.get_value()
         if len(vstr) > 6:
             font.setPixelSize(3.5 * self._scale_factor)
         elif len(vstr) > 3: