+++ python/lsst/ts/criopy/gui/sal/topic_window.py
@@ -34,28 +34,29 @@ from PySide6.QtWidgets import (
 
 from ...salcomm import MetaSAL
 from ..actuatorsdisplay import ForceActuatorItem
-from ..custom_labels import DockWindow, WarningLabel
+from ..custom_labels import WarningLabel
 from .time_delta_label import TimeDeltaLabel
 from .topic_collection import TopicCollection
 from .topic_data import TopicField
 
 
-class TopicWindow(DockWindow):
+class TopicWindow(QSplitter):
     """
     Abstract class for widget and graphics display of selected M1M3 values.
     Children classes must implement update_values(data) method.
 
     Parameters
     ----------
-    title : str
-        Dock title and name.
-    comm : MetaSAL
+    comm : `MetaSAL`
         SAL instance to communicate with SAL.
-    collection : TopicCollection
+    collection : `TopicCollection`
         Collections of data associated with widget.
-    user_widget : QWidget
+    user_widget : `QWidget`
         Widget to be displayed on left from value selection. Its content shall
         be update in update_values(data) method.
+    detail_widget : `QWidget`, optional
+        Widget displaying details of selected actuator. If not provided, basic
+        selected actuator data will be displayed.
 
     Methods
     -------
@@ -66,7 +67,7 @@ class TopicWindow(DockWindow):
 
     Attributes
     ----------
-    field : TopicField | None
+    field : `TopicField | None`
         Selected field. Can be used to access details about the field.
     """
 
@@ -74,42 +75,58 @@ class TopicWindow(DockWindow):
 
     def __init__(
         self,
-        title: str,
         comm: MetaSAL,
         collection: TopicCollection,
         user_widget: QWidget,
+        detail_widget: QWidget | None = None,
     ):
-        super().__init__(title)
+        super().__init__()
 
         self.comm = comm
         self.collection = collection
+        self.user_widget = user_widget
 
-        splitter = QSplitter()
+        self.field: TopicField | None = None
 
         plot_layout = QVBoxLayout()
         selection_layout = QVBoxLayout()
-        details_layout = QFormLayout()
         filter_layout = QHBoxLayout()
 
         w_left = QWidget()
         w_left.setLayout(plot_layout)
-        splitter.addWidget(w_left)
+        self.addWidget(w_left)
         w_right = QWidget()
         w_right.setLayout(selection_layout)
-        splitter.addWidget(w_right)
+        self.addWidget(w_right)
 
-        splitter.setCollapsible(0, False)
-        splitter.setStretchFactor(0, 10)
-        splitter.setStretchFactor(1, 0)
+        self.setCollapsible(0, False)
+        self.setStretchFactor(0, 10)
+        self.setStretchFactor(1, 0)
 
-        selection_layout.addLayout(details_layout)
+        if detail_widget is not None:
+            selection_layout.addWidget(detail_widget)
+        else:
+            self.selected_actuator_id_label = QLabel()
+            self.selected_actuator_value_label = QLabel()
+            self.selected_actuator_warning_label = WarningLabel()
+            self.last_updated_label = TimeDeltaLabel()
+
+            details_layout = QFormLayout()
+            details_layout.addRow(QLabel("Selected Actuator Details"), QLabel(""))
+            details_layout.addRow(
+                QLabel("Actuator Id"), self.selected_actuator_id_label
+            )
+            details_layout.addRow(
+                QLabel("Actuator Value"), self.selected_actuator_value_label
+            )
+            details_layout.addRow(
+                QLabel("Actuator Warning"), self.selected_actuator_warning_label
+            )
+            details_layout.addRow(QLabel("Last Updated"), self.last_updated_label)
+            selection_layout.addLayout(details_layout)
+
         selection_layout.addWidget(QLabel("Filter Data"))
         selection_layout.addLayout(filter_layout)
-
-        self.selected_actuator_id_label = QLabel()
-        self.selected_actuator_value_label = QLabel()
-        self.selected_actuator_warning_label = WarningLabel()
-        self.last_updated_label = TimeDeltaLabel()
 
         self.topic_list = QListWidget()
         self.topic_list.currentRowChanged.connect(self.current_topic_changed)
@@ -118,24 +135,12 @@ class TopicWindow(DockWindow):
         self.field_list = QListWidget()
         self.field_list.currentRowChanged.connect(self.current_field_changed)
 
-        plot_layout.addWidget(user_widget)
-
-        details_layout.addRow(QLabel("Selected Actuator Details"), QLabel(""))
-        details_layout.addRow(QLabel("Actuator Id"), self.selected_actuator_id_label)
-        details_layout.addRow(
-            QLabel("Actuator Value"), self.selected_actuator_value_label
-        )
-        details_layout.addRow(
-            QLabel("Actuator Warning"), self.selected_actuator_warning_label
-        )
-        details_layout.addRow(QLabel("Last Updated"), self.last_updated_label)
+        plot_layout.addWidget(self.user_widget)
 
         filter_layout.addWidget(self.topic_list)
         filter_layout.addWidget(self.field_list)
 
         self.topic_list.setCurrentRow(0)
-
-        self.setWidget(splitter)
 
     @Slot()
     def current_topic_changed(self, topic_index: int) -> None:
@@ -147,7 +152,7 @@ class TopicWindow(DockWindow):
         for field in self.collection.topics[topic_index].fields:
             self.field_list.addItem(field.name)
 
-        field_index = self.collection.topics[topic_index].selectedField
+        field_index = self.collection.topics[topic_index].selected_field
         if field_index < 0:
             self._set_unknown()
             return
@@ -162,7 +167,18 @@ class TopicWindow(DockWindow):
             self._set_unknown()
             return
         self._change_field(topic_index, field_index)
-        self.collection.topics[topic_index].selectedField = field_index
+        self.collection.topics[topic_index].selected_field = field_index
+
+    def field_changed(self, field: TopicField) -> None:
+        """
+        May be overwritten in child classes to setup scaling etc.
+
+        Parameters
+        ----------
+        field : `TopicField`
+            New topic.
+        """
+        pass
 
     def _set_unknown(self) -> None:
         self.last_updated_label.set_unknown()
@@ -173,7 +189,6 @@ class TopicWindow(DockWindow):
 
         Parameters
         ----------
-
         selected_actuator : `ForceActuatorItem`
             Contains actuator ID (selected actuator ID), data (selected
             actuator current value) and warning (boolean, true if value is in
@@ -191,12 +206,15 @@ class TopicWindow(DockWindow):
 
     def _change_field(self, topic_index: int, field_index: int) -> None:
         """
-        Redraw actuators with new values.
+        Redraw actuators with new values fields.
+
+        Para
         """
         topic = self.collection.topics[topic_index]
         self.field = topic.fields[field_index]
         try:
             self.collection.change_topic(topic_index, self.data_changed, self.comm)
+            self.field_changed(self.field)
 
             data = getattr(self.comm.remote, topic.getTopic()).get()
             self.data_changed(data)
@@ -214,9 +232,12 @@ class TopicWindow(DockWindow):
 
         Parameters
         ----------
-        data : BaseMsgType
+        data : `BaseMsgType`
             New values, retrieved from SAL.
         """
+        if self.field is None:
+            raise RuntimeError("update_values was called with empty field.")
+
         self.update_values(data)
         if data is None:
             self._set_unknown()