+++ python/lsst/ts/criopy/m1m3ts/thermal_data.py
@@ -39,7 +39,11 @@ class Thermals(TopicCollection):
                         "Heater PWM", "heaterPWM", self.THERMAL_INDEX, fmt="{v:.2f}%"
                     ),
                     TopicField(
-                        "Fan RPM", "fanRPM", self.THERMAL_INDEX, fmt="{v:d} RPM"
+                        "Fan RPM",
+                        "fanRPM",
+                        self.THERMAL_INDEX,
+                        Scales.INTEGER,
+                        fmt="{v:d} RPM",
                     ),
                 ],
                 "fcuTargets",
@@ -54,13 +58,19 @@ class Thermals(TopicCollection):
                         self.THERMAL_INDEX,
                         fmt="{v:.3f} °C",
                     ),
-                    TopicField("Fan RPM", "fanRPM", self.THERMAL_INDEX),
+                    TopicField("Fan RPM", "fanRPM", self.THERMAL_INDEX, Scales.INTEGER),
                     TopicField(
                         "Absolute temperature",
                         "absoluteTemperature",
                         self.THERMAL_INDEX,
                         fmt="{v:.3f} °C",
                     ),
+                    WarningField("ILC Fault", "ilcFault", self.THERMAL_INDEX),
+                    WarningField(
+                        "Heater Disabled", "heaterDisabled", self.THERMAL_INDEX
+                    ),
+                    WarningField("Heater Breaker", "heaterBreaker", self.THERMAL_INDEX),
+                    WarningField("Fan Breaker", "fanBreaker", self.THERMAL_INDEX),
                 ],
                 "thermalData",
                 False,