+++ python/lsst/ts/criopy/m1m3ts/thermal_value_page_widget.py
@@ -19,9 +19,12 @@
 
 import enum
 import typing
+from math import isnan
 
 from lsst.ts.salobj import BaseMsgType
-from PySide6.QtCore import Slot
+from lsst.ts.xml.tables.m1m3 import FCUTable
+from PySide6.QtCore import Qt, Slot
+from PySide6.QtGui import QBrush, QColor
 from PySide6.QtWidgets import (
     QAbstractItemView,
     QGridLayout,
@@ -45,36 +48,55 @@ class Buttons(enum.IntEnum):
     HEATERS = 2
 
 
+class ByteSpin(QSpinBox):
+    def __init__(self) -> None:
+        super().__init__()
+        self.setRange(0, 255)
+
+
 class DataWidget(QTableWidget):
     """Table with ILC values. Stores ILC values."""
 
     def __init__(self) -> None:
         super().__init__(10, 10)
-        for r in range(0, 10):
-            self.setVerticalHeaderItem(r, QTableWidgetItem(str(r * 10)))
-            for c in range(0, 10):
-                item = QTableWidgetItem("")
-                self.setItem(r, c, item)
+        for c in range(10):
+            self.setVerticalHeaderItem(c, QTableWidgetItem(str(c * 10)))
+        for index in range(96):
+            item = QTableWidgetItem("---")
+            self.setItem(int(index / 10), index % 10, item)
         self.empty()
         self.setEditTriggers(QAbstractItemView.NoEditTriggers)
 
     def empty(self) -> None:
-        self.setValues(range(1, 97))
+        self.set_values(range(1, 97))
+        self.reset_background()
 
-    def setValues(self, data: BaseMsgType, fmt: str | None = None) -> None:
+    def reset_background(self) -> None:
+        d_b = self.palette().base()
+        for index in range(96):
+            self.get_item(index).setBackground(d_b)
+
+    def get_item(self, index: int) -> QTableWidgetItem:
+        return self.item(int(index / 10), index % 10)
+
+    def set_value(self, index: int, value: str) -> None:
+        self.get_item(index).setText(str(value))
+
+    def set_values(self, data: BaseMsgType, fmt: str | None = None) -> None:
         r = 0
         c = 0
         for value in data:
+            item = self.item(r, c)
             if fmt is None:
-                self.item(r, c).setText(str(value))
+                item.setText(str(value))
             else:
-                self.item(r, c).setText(fmt.format(v=value))
+                item.setText(fmt.format(v=value))
             c += 1
             if c >= self.columnCount():
                 c = 0
                 r += 1
 
-    def getValues(self) -> list[int]:
+    def get_values(self) -> list[int]:
         data: list[int] = []
         for r in range(0, 10):
             for c in range(0, 10):
@@ -82,6 +104,11 @@ class DataWidget(QTableWidget):
                 if index < 96:
                     data.append(int(self.item(r, c).text()))
         return data
+
+    def mask_backround(self, mask: list[bool], brush: QBrush) -> None:
+        for index, masked in enumerate(mask):
+            if masked:
+                self.get_item(index).setBackground(brush)
 
 
 class CommandWidget(QWidget):
@@ -161,13 +188,24 @@ class CommandWidget(QWidget):
         self.set_heaters_button = self.SetButton(self, Buttons.HEATERS)
         self.set_fans_button = self.SetButton(self, Buttons.FANS)
 
-        self.flat_demand = QSpinBox()
-        self.flat_demand.setRange(0, 255)
+        self.flat_demand = ByteSpin()
+        self.m1_demand = ByteSpin()
+        self.m3_demand = ByteSpin()
 
         self.set_constant_button = QPushButton("Set constant")
         self.set_constant_button.setToolTip("Sets all target values to given constant")
         self.set_constant_button.setDisabled(True)
         self.set_constant_button.clicked.connect(self.set_constant)
+
+        self.set_m1_button = QPushButton("Set M1")
+        self.set_m1_button.setToolTip("Sets M1 FCUs values to given value")
+        self.set_m1_button.setDisabled(True)
+        self.set_m1_button.clicked.connect(self.set_m1)
+
+        self.set_m3_button = QPushButton("Set M3")
+        self.set_m3_button.setToolTip("Sets M3 FCUs values to given value")
+        self.set_m3_button.setDisabled(True)
+        self.set_m3_button.clicked.connect(self.set_m3)
 
         self.cancel_button = QPushButton("Cancel")
         self.cancel_button.setToolTip("Cancel value editing")
@@ -180,6 +218,11 @@ class CommandWidget(QWidget):
         command_layout.addWidget(self.cancel_button, 0, 2, 2, 1)
         command_layout.addWidget(self.set_heaters_button, 1, 0)
         command_layout.addWidget(self.set_fans_button, 1, 1)
+
+        command_layout.addWidget(self.m1_demand, 0, 4)
+        command_layout.addWidget(self.m3_demand, 1, 4)
+        command_layout.addWidget(self.set_m1_button, 0, 5)
+        command_layout.addWidget(self.set_m3_button, 1, 5)
 
         hBox = QHBoxLayout()
         hBox.addLayout(command_layout)
@@ -194,13 +237,27 @@ class CommandWidget(QWidget):
 
     @Slot()
     def set_constant(self) -> None:
-        self.data_widget.setValues([self.flat_demand.text()] * 96)
+        self.data_widget.set_values([self.flat_demand.text()] * 96)
+
+    @Slot()
+    def set_m1(self) -> None:
+        value = self.m1_demand.text()
+        for fcu in [fcu for fcu in FCUTable if fcu.is_m1()]:
+            self.data_widget.set_value(fcu.index, value)
+
+    @Slot()
+    def set_m3(self) -> None:
+        value = self.m3_demand.text()
+        for fcu in [fcu for fcu in FCUTable if fcu.is_m3()]:
+            self.data_widget.set_value(fcu.index, value)
 
     @Slot()
     def cancel(self) -> None:
         self.freezed = False
         self.data_widget.setEditTriggers(QAbstractItemView.NoEditTriggers)
         self.set_constant_button.setDisabled(True)
+        self.set_m1_button.setDisabled(True)
+        self.set_m3_button.setDisabled(True)
         self.set_fans_button.setEnabled(True)
         self.set_fans_button.cancel()
         self.set_heaters_button.setEnabled(True)
@@ -222,9 +279,11 @@ class CommandWidget(QWidget):
         self.data_widget.setEnabled(True)
         self.cancel_button.setEnabled(True)
         self.set_constant_button.setEnabled(True)
+        self.set_m1_button.setEnabled(True)
+        self.set_m3_button.setEnabled(True)
 
     async def heater_fan_demand(self, kind: Buttons) -> None:
-        data = self.data_widget.getValues()
+        data = self.data_widget.get_values()
         if kind == Buttons.HEATERS:
             await self._heater_fan_demand(heaterPWM=data, fanRPM=self.fans)
             self.heaters = data
@@ -235,7 +294,7 @@ class CommandWidget(QWidget):
         self.cancel()
 
     def update_values(
-        self, values: typing.Any, freeze: bool = False, fmt: str | None = None
+        self, values: None | BaseMsgType, freeze: bool = False, fmt: str | None = None
     ) -> None:
         if self.freezed:
             return
@@ -246,7 +305,10 @@ class CommandWidget(QWidget):
             self.data_widget.empty()
             return
 
-        self.data_widget.setValues(values, fmt=fmt)
+        self.data_widget.set_values(values, fmt=fmt)
+
+    def mask_backround(self, mask: list[bool], background: QBrush) -> None:
+        self.data_widget.mask_backround(mask, background)
 
 
 class ThermalValuePageWidget(TopicWindow):
@@ -261,12 +323,54 @@ class ThermalValuePageWidget(TopicWindow):
     def __init__(self, m1m3ts: MetaSAL):
         self.command_widget = CommandWidget(m1m3ts)
 
-        super().__init__("Thermal Values", m1m3ts, Thermals(), self.command_widget)
+        super().__init__(m1m3ts, Thermals(), self.command_widget)
 
     def update_values(self, data: BaseMsgType) -> None:
         if data is None or self.field is None:
             self.command_widget.update_values(None)
         else:
-            self.command_widget.update_values(
-                self.field.getValue(data), fmt=self.field.fmt
-            )
+            values = self.field.getValue(data)
+            self.command_widget.data_widget.reset_background()
+            self.command_widget.update_values(values, fmt=self.field.fmt)
+            if self.field.field_name in ["heaterPWM"]:
+                self.command_widget.mask_backround(
+                    [heater == 0 for heater in values], Qt.red
+                )
+                self.command_widget.mask_backround(
+                    [heater == 100 for heater in values], Qt.blue
+                )
+            if self.field.field_name in "absoluteTemperature":
+                self.command_widget.mask_backround(
+                    [isnan(t) for t in values], Qt.darkRed
+                )
+                target = self.comm.remote.evt_appliedSetpoints.get()
+                if target is not None and not isnan(target.heatersSetpoint):
+                    tt = target.heatersSetpoint
+                    t_diff = [t - tt for t in values]
+                    self.command_widget.mask_backround(
+                        [td < -0.025 for td in t_diff], QColor("#99CCFF")
+                    )
+                    self.command_widget.mask_backround(
+                        [td > 0.025 for td in t_diff], QColor("#FF9999")
+                    )
+                    self.command_widget.mask_backround(
+                        [td < -0.05 for td in t_diff], Qt.blue
+                    )
+                    self.command_widget.mask_backround(
+                        [td > 0.05 for td in t_diff], Qt.red
+                    )
+            if self.field.field_name in ["heaterPWM", "fanRPM", "absoluteTemperature"]:
+                t_w = self.comm.remote.evt_thermalWarning.get()
+                if t_w is not None:
+                    self.command_widget.mask_backround(
+                        t_w.breakerHeater1Error, QBrush(Qt.red, Qt.Dense6Pattern)
+                    )
+            elif isinstance(values[0], bool):
+                self.command_widget.mask_backround(
+                    values, QBrush(Qt.red, Qt.Dense3Pattern)
+                )
+            thermal_settings = self.comm.remote.evt_thermalSettings.get()
+            if thermal_settings is not None:
+                self.command_widget.mask_backround(
+                    [not (d) for d in thermal_settings.enabledFCU], QBrush(Qt.gray)
+                )